+++ src/yafowil/widget/cron/widget.py
@@ -1,13 +1,17 @@
 # -*- coding: utf-8 -*-
 from node.utils import UNSET
+from yafowil.base import ExtractionError
 from yafowil.base import factory
 from yafowil.base import fetch_value
+from yafowil.common import generic_required_extractor
 from yafowil.compound import compound_extractor
 from yafowil.compound import compound_renderer
 from yafowil.tsf import TSF
 from yafowil.utils import attr_value
 from yafowil.utils import cssclasses
 from yafowil.utils import cssid
+from yafowil.utils import EMPTY_VALUE
+from yafowil.utils import managedprops
 
 
 _ = TSF('yafowil.widget.cron')
@@ -31,29 +35,41 @@ factory.register(
 factory.doc['blueprint']['cron_value_edit_action'] = UNSET
 
 
+@managedprops('emptyvalue')
 def cron_extractor(widget, data):
     # instanciate subwidgets
     widget()
     # extract subwidgets
     compound_extractor(widget, data)
-    # XXX: empty value and required handling
     minute = data['minute']
-    if not minute.extracted:
-        return minute.extracted
-    value = '{0} {1} {2} {3} {4} {5}'.format(
-        data['minute'].extracted,
-        data['hour'].extracted,
-        data['dom'].extracted,
-        data['month'].extracted,
-        data['dow'].extracted,
-        data['year'].extracted
-    )
-    return value
+    # if one subwidget is UNSET, whole widget not found on request
+    if minute.extracted is UNSET:
+        return UNSET
+    minute = data['minute'].extracted
+    hour = data['hour'].extracted
+    dom = data['dom'].extracted
+    month = data['month'].extracted
+    dow = data['dow'].extracted
+    year = data['year'].extracted
+    # if all values missing, we have no cron rule
+    if not minute and not hour and not dom \
+            and not month and not dow and not year:
+        return attr_value('emptyvalue', widget, data, EMPTY_VALUE)
+    # if one value missing, we have an invalid cron rule
+    if not (minute and hour and dom and month and dow and year):
+        raise ExtractionError(_(
+            'invalid_cron_rule',
+            default=(
+                'Invalid cron rule. You must at least '
+                'select one item for each criteria'
+            )
+        ))
+    return '{0} {1} {2} {3} {4} {5}'.format(minute, hour, dom, month, dow, year)
 
 
 def cron_edit_renderer(widget, data):
     value = fetch_value(widget, data)
-    if value is not UNSET:
+    if value is not UNSET and value is not attr_value('emptyvalue', widget, data, EMPTY_VALUE):
         value = [it.strip() for it in value.split(' ') if it.strip()]
         if len(value) == 5:
             value.append('*')
@@ -143,7 +159,8 @@ def cron_display_renderer(widget, data):
 factory.register(
     'cron',
     extractors=[
-        cron_extractor
+        cron_extractor,
+        generic_required_extractor
     ],
     edit_renderers=[
         cron_edit_renderer,
@@ -154,7 +171,7 @@ factory.register(
     ]
 )
 
-factory.doc['blueprint']['cron'] ="""\
+factory.doc['blueprint']['cron'] = """\
 Add-on blueprint
 `yafowil.widget.cron <http://github.com/bluedynamics/yafowil.widget.cron/>`_ .
 """
@@ -177,4 +194,9 @@ Start year. Defaults to current year.
 factory.defaults['cron.end_year'] = None
 factory.doc['props']['cron.lang'] = """\
 End year. Defaults to current year + 9
+"""
+
+factory.doc['props']['cron.emptyvalue'] = """\
+If cron rule value empty, return as extracted
+value.
 """